ACCUMULO-4780 Add overflow check to seq num in CommitSession (#360)
Add overflow check to sequence number in CommitSession.

Ticket suggested adding overflow check to sequence number or converting to long.
This PR converted the sequence number from and int to long to prevent fear of future
overflow. The change propagated to several other files that made use of or called
methods that utilized the sequence number.

Modified the various files and then re-ran tests and integration tests to verify nothing
had broken. All tests passes.

+++ server/tserver/src/main/java/org/apache/accumulo/tserver/TabletMutations.java
@@ -23,11 +23,11 @@ import org.apache.accumulo.core.data.Mutation;
 
 public class TabletMutations {
   private final int tid;
-  private final int seq;
+  private final long seq;
   private final List<Mutation> mutations;
   private final Durability durability;
 
-  public TabletMutations(int tid, int seq, List<Mutation> mutations, Durability durability) {
+  public TabletMutations(int tid, long seq, List<Mutation> mutations, Durability durability) {
     this.tid = tid;
     this.seq = seq;
     this.mutations = mutations;
@@ -42,7 +42,7 @@ public class TabletMutations {
     return tid;
   }
 
-  public int getSeq() {
+  public long getSeq() {
     return seq;
   }
